Fix some issues with the public api service

+++ frontend/lib/models/nutriment_model.dart
@@ -2,7 +2,7 @@ class Nutriments {
   int? fat;
   int? proteins;
   int? salt;
-  double? sodium;
+  int? sodium;
   int? sugars;
   int? fiber;
   int? saturatedFat;
